+++ service/server.ts
@@ -1,6 +1,6 @@
 import express from "express"
 import graphqlHTTP from "express-graphql"
-import requestDevice from "../src/nodeusb"
+import requestDevice from "../src/nodewebusb"
 import { createUSBBus } from "../src/usb"
 import { getSchema, rootValue } from "../src/graphql"
